Split interfaces
